+++ app/src/main/java/nie/translator/rtranslatordevedition/voice_translation/_walkie_talkie_mode/_walkie_talkie/WalkieTalkieService.java
@@ -75,6 +75,9 @@ public class WalkieTalkieService extends VoiceTranslationService {
     }
 
     @Override public int onStartCommand(Intent intent, int flags, int startId) {
+        if (intent == null) {
+            return super.onStartCommand(null, flags, startId);
+        }
         applyLanguages((CustomLocale) intent.getSerializableExtra("firstLanguage"), (CustomLocale) intent.getSerializableExtra("secondLanguage"));
         return super.onStartCommand(intent, flags, startId);
     }